fix topbar hmr
